Im not sure if this warrants a new version, but late move reduction has been implemented resulting in about a 90 elo boost of performance
commented out code is left in as experimental

note that I believe the king friendly pawns and enemy pawns are somewhat bugged in the feature generator, or at least weighted erroneously.

things that need to be refined/experimented with further:
-null move pruning
- futility pruning
- aspiration windows

Last thing to try and add:
delta pruning in QS search. REASON: currently the majority of nodes are in this search, and we have 0 heuristics to speed it up outside of alpha beta.