this is more or less an empty commit, but wanted to note some things
I have watched HOURS of this engine play, and as the depth gets more and more competent, something is abundantly clear:
-pawns dont get pushed in the endgame, even passers
- king pawns get pushed WAY too willy nilly.

hypothesis: Somethings awry in the feature generation or scoring for the features that have to do with pawn in the endgame, as well as friendly pawn. Could also potentially be an issue with flipping the board.