Ga is up and running
Gotten the best results running the ga on only quiet positions evaluated with stockfish for one second. The numbers look promosing but some of the finer tuned values are suspect. My guess is it could be a result of the piece tables. My next try is to include the piece tables. From there, I will work on the quality of the dataset, and then optimizing the hyperparameters of the GA