this is a huge discovery, the engine is not worse if you get rid of piece tables
There are many things in chess programming that are considered approximations. One of those is static piece tables. I hate them, they have no ability to know about the relationships between pieces, which is what chess is all about. I just found out that they are not really helping in any phase but the opening, in which we can just use an opening book to get our pieces on reasonable squares...